Spark on DLWorkspace (#323)

+++ src/docker-images/hdfs/bootstrap_hdfs.py
@@ -73,9 +73,16 @@ datanode:    Launch datanode.
 #           f.close()
 #           print logging_config
 #           logutils.dictconfig.dictConfig(logging_config)
-        utils.render_template("hdfs-site.xml.in-docker", "/usr/local/hadoop/etc/hadoop/hdfs-site.xml",config, verbose=verbose)
+        isHA = "namenode2" in config["namenode"]
+        if isHA:
+            utils.render_template("hdfs-site.xml.in-docker", "/usr/local/hadoop/etc/hadoop/hdfs-site.xml",config, verbose=verbose)
+        else:
+            utils.render_template("hdfs-site-single.xml.in-docker", "/usr/local/hadoop/etc/hadoop/hdfs-site.xml",config, verbose=verbose)
         utils.render_template("mapred-site.xml.in-docker", "/usr/local/hadoop/etc/hadoop/mapred-site.xml",config, verbose=verbose)
-        utils.render_template("yarn-site.xml.in-docker", "/usr/local/hadoop/etc/hadoop/yarn-site.xml",config, verbose=verbose)
+        if isHA:
+            utils.render_template("yarn-site.xml.in-docker", "/usr/local/hadoop/etc/hadoop/yarn-site.xml",config, verbose=verbose)
+        else:
+            utils.render_template("yarn-site-single.xml.in-docker", "/usr/local/hadoop/etc/hadoop/yarn-site.xml",config, verbose=verbose)
     except Exception as e:
         print "boostrap_hdfs.py fails during initialization, exception %s" % e
         exit()
@@ -106,14 +113,16 @@ datanode:    Launch datanode.
         force = "" if not args.force else "-force "
         cmd = "/usr/local/hadoop/bin/hdfs namenode -format %s -nonInteractive" % force
         exec_with_output( cmd )
-        cmd = "/usr/local/hadoop/bin/hdfs zkfc -formatZK -nonInteractive"
-        exec_with_output( cmd )
+        if isHA:
+            cmd = "/usr/local/hadoop/bin/hdfs zkfc -formatZK -nonInteractive"
+            exec_with_output( cmd )
         print "Format namenode and zkfc ..... "
-        remotedir = os.path.join( config["namenode"]["data"], "current")
-        localdir = os.path.join( config["namenode"]["localdata"], "current")
-        exec_with_output( "rm -rf %s" % remotedir )
-        exec_with_output( "cp -r %s %s" % ( localdir, remotedir ) )
-        print "Copy data from %s to %s" % ( localdir, remotedir )
+        if isHA:
+            remotedir = os.path.join( config["namenode"]["data"], "current")
+            localdir = os.path.join( config["namenode"]["localdata"], "current")
+            exec_with_output( "rm -rf %s" % remotedir )
+            exec_with_output( "cp -r %s %s" % ( localdir, remotedir ) )
+            print "Copy data from %s to %s" % ( localdir, remotedir )
     elif server == "copy":
         remotedir = os.path.join( config["namenode"]["data"], "current")
         localdir = os.path.join( config["namenode"]["localdata"], "current")
